requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
distribute==0.6.24
gevent==1.0.1
greenlet==0.4.2
itsdangerous==0.24
pymongo==2.7.1
pyserial==2.7
redis==2.10.1
wsgiref==0.1.2
